+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## [2.0.0] - 2024-12-18
+
+- Separate table for storing chart constant
+- Use Qman's script to get all score records
+- Generate Best Songs Image and send to Discord
+
 ## [1.1.0] - 2024-09-01
 
 - Add Column: Player Level, Name, Team Name and Honor Text (称号)